+++ pnmlFw-Tests/test/fr/lip6/move/pnml/hlpn/integers/hlapi/NumberConstantHLAPITest.java
@@ -11,7 +11,7 @@
  *    Bastien Bouzerau (UPMC) - Architecture 
  *    Guillaume Giffo (UPMC) - Code generation refactoring, High-level API
  *
- * $Id ggiffo, Wed Feb 10 14:59:13 CET 2016$
+ * $Id ggiffo, Thu Feb 11 16:29:58 CET 2016$
  */
 package fr.lip6.move.pnml.hlpn.integers.hlapi;
 
@@ -184,7 +184,7 @@ public class NumberConstantHLAPITest {
 		
 	private HLPNNumberHLAPI itemtype;
 		
-	private Integer itemvalue;
+	private Long itemvalue;
 		
 	
 	private OperatorHLAPI itemcontainerOperator;
@@ -237,7 +237,7 @@ public class NumberConstantHLAPITest {
 				
 			
 		
-			itemvalue = new Integer("0");
+			itemvalue = new Long("0");
