+++ pnmlFw-Tests/test/fr/lip6/move/pnml/symmetricnet/integers/hlapi/NumberConstantHLAPITest.java
@@ -11,7 +11,7 @@
  *    Bastien Bouzerau (UPMC) - Architecture 
  *    Guillaume Giffo (UPMC) - Code generation refactoring, High-level API
  *
- * $Id ggiffo, Wed Feb 10 14:59:12 CET 2016$
+ * $Id ggiffo, Thu Feb 11 16:29:02 CET 2016$
  */
 package fr.lip6.move.pnml.symmetricnet.integers.hlapi;
 
@@ -166,7 +166,7 @@ public class NumberConstantHLAPITest {
 		
 	private HLPNNumberHLAPI itemtype;
 		
-	private Integer itemvalue;
+	private Long itemvalue;
 		
 	
 	private OperatorHLAPI itemcontainerOperator;
@@ -219,7 +219,7 @@ public class NumberConstantHLAPITest {
 				
 			
 		
-			itemvalue = new Integer("0");
+			itemvalue = new Long("0");
